Stop triggering a prove for .# files
Emacs uses a symlink called `.#filename` as a lockfile to provide safety against multi-user editing. I don't need provewatcher to trigger on these requests. 

Alternatively, `provewatcher` could accept an --ignore option as a regex or something and use it in place of this to filter files more generally, if you'd prefer me to make that change.